operator/requirements.txt
Flask
flask-cors
python-dotenv
gunicorn
livekit-api
